Fix all Vortice.Vulkan 3.2.1 compilation errors
- VulkanUtil.cs: fix vkEnumerateInstanceLayerProperties/ExtensionProperties to use pointer overloads; add VkDeviceApi param to TransitionImageLayout
- VkTexture.cs: remove GetImageMemoryRequirements2/GetBufferMemoryRequirements2 delegate pattern, use DeviceApi directly; pass gd.DeviceApi to TransitionImageLayout
- VkBuffer.cs: remove GetBufferMemoryRequirements2 delegate pattern, use DeviceApi directly
- VkCommandList.cs: fix VkRect2D ctors; fix ref->pointer for barriers/regions; add VkDeviceApi param to CopyTextureCore_VkCommandBuffer; fix VkImageBlit initializer; fix VkFragmentShadingRateCombinerOpKHR array type; fix vkAllocateCommandBuffers out->pointer
- VkDescriptorPoolManager.cs: replace .New() ctors; fix vkAllocateDescriptorSets out->pointer; fix PoolInfo.Free to take VkDeviceApi
- VkGraphicsDevice.cs: fix vkGetInstanceProcAddr call; fix VkVersion ctor; fix duplicate props variable; fix deviceName.Length->256; fix vkCreateDevice &device->local; fix vkAllocateCommandBuffers out->local; pass DeviceApi to CopyTextureCore
- VkPipeline.cs: fix vkCreateGraphicsPipelines/vkCreateComputePipelines out->pointer with fully-qualified VkPipeline type
- VkSurfaceInfo.cs: cast ulong window to IntPtr for XlibSwapchainSource
- VkSurfaceUtil.cs: cast void* pLayer to nint
- VkSwapchain.cs: fix managed array address-of with fixed blocks; fix vkGetPhysicalDeviceSurfaceSupportKHR out->pointer
- VkSwapchainFramebuffer.cs: fix vkGetSwapchainImagesKHR out->fixed block
- VkTextureUpdateBatch.cs: fix ref region->pointer in vkCmdCopyBufferToImage

+++ src/Veldrid/Vk/VkSurfaceInfo.cs
@@ -81,7 +81,7 @@ namespace Veldrid.Vk
 
         internal override SwapchainSource GetSurfaceSource()
         {
-            return new XlibSwapchainSource(display, window);
+            return new XlibSwapchainSource(display, (IntPtr)(long)window);
         }
     }
 }
